refactor: rename util to utils close #422
