+++ packages/devui-vue/devui/modal/index.ts
@@ -4,7 +4,7 @@ import Header from './src/components/header';
 import Body from './src/components/body';
 import Footer from './src/components/footer';
 import { ModalService } from './src/services/modal-service';
-import { inBrowser } from '../shared/util/common-var';
+import { inBrowser } from '../shared/utils/common-var';
 
 export * from './src/modal-types';
