+++ packages/devui-vue/devui/overlay/index.ts
@@ -1,7 +1,7 @@
 import type { App } from 'vue';
 import { FixedOverlay } from './src/fixed-overlay';
 import { FlexibleOverlay } from './src/flexible-overlay';
-import { inBrowser } from '../shared/util/common-var';
+import { inBrowser } from '../shared/utils/common-var';
 
 export * from './src/fixed-overlay/fixed-overlay-types';
 export * from './src/flexible-overlay/flexible-overlay-types';
