+++ packages/devui-vue/devui/shared/utils/common-var.ts
