+++ packages/devui-vue/devui/tree/src/core/utils.ts
@@ -1,4 +1,4 @@
-import { randomId } from '../../../shared/util';
+import { randomId } from '../../../shared/utils';
 import { IInnerTreeNode, ITreeNode } from './use-tree-types';
 
 export function flatToNested(flatTree: IInnerTreeNode[]): ITreeNode[] {
